refactor: epoch as install-signature replay guard, not actor stamp
Reframes the local-channel epoch so it guards install/change *signatures*
at the apply path rather than stamping live actors:

- Drop ActorConfig.installEpoch and the per-authentication epoch check
  (removes ActorEpochRevoked and the extra SLOAD for non-admin actors).
  A bump now cancels pending signatures, not live agents.
- Replace the applySignedEpochBump entry point with a BUMP_EPOCH (0x03)
  change type inside applySignedActorChanges, so [REVOKE_ACTOR(x),
  BUMP_EPOCH] is one atomic signature: x is revoked and its install
  signature is permanently dead. Local channel only, empty data, never on
  the MAX-sentinel path.
- Keep expiry != 0 required on MAX-sentinel installs.
- Widen AccountState.lockUnion uint40 -> uint48 so lock unlocksAt matches
  the actor/expiry timestamp width (both Unix seconds); 1 reserved byte.

+++ test/unit/AccountConfiguration/createAccount.t.sol
@@ -268,7 +268,7 @@ contract CreateAccountTest is AccountConfigurationTest {
         assertEq(accountConfiguration.getChangeSequences(account).local, 1);
         assertEq(accountConfiguration.getChangeSequences(account).multichain, 0);
 
-        (bool locked, bool hasInitiatedUnlock, uint40 unlocksAt, uint16 unlockDelay) =
+        (bool locked, bool hasInitiatedUnlock, uint48 unlocksAt, uint16 unlockDelay) =
             accountConfiguration.getLockStatus(account);
         assertFalse(locked);
         assertFalse(hasInitiatedUnlock);
